# api/gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package api

import (
	"github.com/gin-gonic/gin"
)

// CalculationRequestNumbers defines model for CalculationRequestNumbers.
type CalculationRequestNumbers struct {
	Number1 int `json:"number1"`
	Number2 int `json:"number2"`
}

// DivisionRequest defines model for DivisionRequest.
type DivisionRequest struct {
	Dividend int `json:"dividend"`
	Divisor  int `json:"divisor"`
}

// Error defines model for Error.
type Error struct {
	Message string `json:"message"`
}

// Result defines model for Result.
type Result struct {
	Result int `json:"result"`
}

// SumRequest An array of numbers to sum
type SumRequest = []int

// AddNumbersJSONRequestBody defines body for AddNumbers for application/json ContentType.
type AddNumbersJSONRequestBody = CalculationRequestNumbers

// DivideNumbersJSONRequestBody defines body for DivideNumbers for application/json ContentType.
type DivideNumbersJSONRequestBody = DivisionRequest

// MultiplyNumbersJSONRequestBody defines body for MultiplyNumbers for application/json ContentType.
type MultiplyNumbersJSONRequestBody = CalculationRequestNumbers

// SubtractNumbersJSONRequestBody defines body for SubtractNumbers for application/json ContentType.
type SubtractNumbersJSONRequestBody = CalculationRequestNumbers

// SumNumbersJSONRequestBody defines body for SumNumbers for application/json ContentType.
type SumNumbersJSONRequestBody = SumRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Add two numbers
	// (POST /add)
	AddNumbers(c *gin.Context)
	// Divide two numbers
	// (POST /divide)
	DivideNumbers(c *gin.Context)
	// Multiply two numbers
	// (POST /multiply)
	MultiplyNumbers(c *gin.Context)
	// Subtract two numbers
	// (POST /subtract)
	SubtractNumbers(c *gin.Context)
	// Add all numbers in an array
	// (POST /sum)
	SumNumbers(c *gin.Context)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// AddNumbers operation middleware
func (siw *ServerInterfaceWrapper) AddNumbers(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.AddNumbers(c)
}

// DivideNumbers operation middleware
func (siw *ServerInterfaceWrapper) DivideNumbers(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DivideNumbers(c)
}

// MultiplyNumbers operation middleware
func (siw *ServerInterfaceWrapper) MultiplyNumbers(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.MultiplyNumbers(c)
}

// SubtractNumbers operation middleware
func (siw *ServerInterfaceWrapper) SubtractNumbers(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.SubtractNumbers(c)
}

// SumNumbers operation middleware
func (siw *ServerInterfaceWrapper) SumNumbers(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.SumNumbers(c)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/add", wrapper.AddNumbers)
	router.POST(options.BaseURL+"/divide", wrapper.DivideNumbers)
	router.POST(options.BaseURL+"/multiply", wrapper.MultiplyNumbers)
	router.POST(options.BaseURL+"/subtract", wrapper.SubtractNumbers)
	router.POST(options.BaseURL+"/sum", wrapper.SumNumbers)
}